requirements.txt
discord.py==1.4.0
python-dotenv==0.14.0